+++ packages/core-dart/test/web_compat/bigint_web_test.dart
@@ -1,4 +1,5 @@
 @TestOn('browser')
+library;
 
 import 'package:bluewhale_core/bluewhale_core.dart';
 import 'package:test/test.dart';